+++ internal/driver/gomobile/canvas_test.go
@@ -158,7 +158,7 @@ func TestCanvas_Dragged(t *testing.T) {
 	c := NewCanvas().(*mobileCanvas)
 	c.SetContent(scroll)
 	c.resize(fyne.NewSize(40, 24))
-	assert.Equal(t, 0, scroll.Offset.Y)
+	assert.Equal(t, float32(0), scroll.Offset.Y)
 
 	c.tapDown(fyne.NewPos(32, 3), 0)
 	c.tapMove(fyne.NewPos(32, 10), 0, func(wid fyne.Draggable, ev *fyne.DragEvent) {
