+++ internal/driver/gomobile/driver.go
@@ -64,7 +64,7 @@ func (d *mobileDriver) currentWindow() fyne.Window {
 	return d.windows[len(d.windows)-1]
 }
 
-func (d *mobileDriver) RenderedTextSize(text string, size int, style fyne.TextStyle) fyne.Size {
+func (d *mobileDriver) RenderedTextSize(text string, size float32, style fyne.TextStyle) fyne.Size {
 	return painter.RenderedTextSize(text, size, style)
 }
 
@@ -145,7 +145,7 @@ func (d *mobileDriver) Run() {
 				}
 
 				if d.freeDirtyTextures(canvas) {
-					newSize := fyne.NewSize(int(float32(currentSize.WidthPx)/canvas.scale), int(float32(currentSize.HeightPx)/canvas.scale))
+					newSize := fyne.NewSize(float32(currentSize.WidthPx)/canvas.scale, float32(currentSize.HeightPx)/canvas.scale)
 
 					if canvas.minSizeChanged() {
 						canvas.ensureMinSize()
